Fixed CardContentRenderer
